Use Correct Network Number in NBP Extended DDP Packets
When broadcasting NBP multicast packets on a non-extended network (LocalTalk), use the network's number in broadcasts. When it is an extended network (EtherTalk), use "0". This mirrors the behavior of commercial AppleTalk routers such as the Shiva Fastpath.

It also avoids annoying behavior with the ImageWriter LocalTalk Option card. That card will unconditionally send back NBP responses to the initial requestor's address (in this case, TashRouter) if using short DDP packets, not the destination in the NBP tuple.