Reject RTMP tuples with network numbers in the startup range
Due to a quirk in netatalk's atalkd router service, it sometimes sends out a single RTMP packet with a tuple in the startup network range. This only happens if atalkd running on two networks and doesn't know the network range of a TashRouter seeded network during startup.

While Inside AppleTalk doesn't rule it out, networks with an active router shouldn't be using any network numbers in the startup range.